Initial CAN driver implementation for SAME51 and major SSA builder fixes.
compiler/ssa:
Implemented support for unsafe.Slice builtin call.
Implemented method calls on const values.
Refactored selector expression addressing algorithm.
Fixed bug with selecting function values from a struct.
Fixed bug with selecting global functions from packages.
Improved method call semantics.
Improved package parsing scheme for tests.
Improved handling of function closures.

runtime:
Correctly wait for synchronization when setting up GCLK, DFLL and DPLL.
Implemented runtime for supporting unsafe.Slice.
Corrected API for slice copy builtin.

peripheral:
Initial CAN implementation for SAME51

encoding/binary:
Partial implementation.

GoIR:
Fixed bug in debug information encoding for unsafe.Pointer.
Improved canonicalization of go.call_go operation to handle packing additional arguments of _func value calls.

other:
Optimized register set/get code to eliminate extra stack allocation.
Implemented panic on abort on Windows for better debugging of internal MLIR/LLVM crashes.

+++ src/runtime/arm/cortexm/scs.go
@@ -107,7 +107,7 @@ func (s *SCS_SHPR1) GetPRI_4() uint8 {
 
 func (s *SCS_SHPR1) SetPRI_4(value uint8) {
 	v := volatile.LoadUint32((*uint32)(s))
-	v &= ^(0xFF << 0)
+	v &= ^uint32(0xFF << 0)
 	v |= uint32(value<<(8-NPRIORITY_BITS)) << 0
 	volatile.StoreUint32((*uint32)(s), v)
 }
@@ -119,7 +119,7 @@ func (s *SCS_SHPR1) GetPRI_5() uint8 {
 
 func (s *SCS_SHPR1) SetPRI_5(value uint8) {
 	v := volatile.LoadUint32((*uint32)(s))
-	v &= ^(0xFF << 8)
+	v &= ^uint32(0xFF << 8)
 	v |= uint32(value<<(8-NPRIORITY_BITS)) << 8
 	volatile.StoreUint32((*uint32)(s), v)
 }
@@ -131,7 +131,7 @@ func (s *SCS_SHPR1) GetPRI_6() uint8 {
 
 func (s *SCS_SHPR1) SetPRI_6(value uint8) {
 	v := volatile.LoadUint32((*uint32)(s))
-	v &= ^(0xFF << 16)
+	v &= ^uint32(0xFF << 16)
 	v |= uint32(value<<(8-NPRIORITY_BITS)) << 16
 	volatile.StoreUint32((*uint32)(s), v)
 }
@@ -143,7 +143,7 @@ func (s *SCS_SHPR2) GetPRI_11() uint8 {
 
 func (s *SCS_SHPR2) SetPRI_11(value uint8) {
 	v := volatile.LoadUint32((*uint32)(s))
-	v &= ^(0xFF << 24)
+	v &= ^uint32(0xFF << 24)
 	v |= uint32(value<<(8-NPRIORITY_BITS)) << 24
 	volatile.StoreUint32((*uint32)(s), v)
 }
@@ -155,7 +155,7 @@ func (s *SCS_SHPR3) GetPRI_12() uint8 {
 
 func (s *SCS_SHPR3) SetPRI_12(value uint8) {
 	v := volatile.LoadUint32((*uint32)(s))
-	v &= ^(0xFF << 0)
+	v &= ^uint32(0xFF << 0)
 	v |= uint32(value<<(8-NPRIORITY_BITS)) << 0
 	volatile.StoreUint32((*uint32)(s), v)
 }
@@ -167,7 +167,7 @@ func (s *SCS_SHPR3) GetPRI_14() uint8 {
 
 func (s *SCS_SHPR3) SetPRI_14(value uint8) {
 	v := volatile.LoadUint32((*uint32)(s))
-	v &= ^(0xFF << 16)
+	v &= ^uint32(0xFF << 16)
 	v |= uint32(value<<(8-NPRIORITY_BITS)) << 16
 	volatile.StoreUint32((*uint32)(s), v)
 }
@@ -179,7 +179,7 @@ func (s *SCS_SHPR3) GetPRI_15() uint8 {
 
 func (s *SCS_SHPR3) SetPRI_15(value uint8) {
 	v := volatile.LoadUint32((*uint32)(s))
-	v &= ^(0xFF << 24)
+	v &= ^uint32(0xFF << 24)
 	v |= uint32(value<<(8-NPRIORITY_BITS)) << 24
 	volatile.StoreUint32((*uint32)(s), v)
 }
